refactor: CONFIG is globalized and config module is minimized into utils
