+++ Cargo.toml
@@ -6,11 +6,10 @@ members = [
   "crates/cli",
   "crates/daemon",
   "crates/report",
-  "crates/config",
   "crates/utils",
   "crates/nixlog",
   "crates/crash-event",
-  "crates/relago-journal-ext", "crates/config",
+  "crates/relago-journal-ext",
 ]
 
 
@@ -37,17 +36,11 @@ toml = { version = "0.8.20", features = ["display"] }
 thiserror = "1"
 anyhow = "1"
 subprocess = "0.2.9"
-sha2 = "0.10"
-hex = "0.4"
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 nom = "8.0.0"
 dbus = "0.9.10"
-dbus-tree = "0.9.2"
 dbus-crossroads = "0.5.3"
-
-# Optimize release binary as much as possible
-[profile.release]
 # codegen-units = 1
 # incremental = true
 # lto = true
