+++ crates/cli/src/main.rs
@@ -1,20 +1,10 @@
 #![allow(unused_must_use)]
 
-use std::{path::PathBuf, process::exit};
-
 use cli::run;
-use config::Config;
+use utils::config::{Config, CONFIG};
 
 fn main() -> anyhow::Result<()> {
-    let CONFIG_FILE: PathBuf = PathBuf::from("config.toml");
-    let mut conf: Config = Config::new();
+    CONFIG.set(|| Config::get_config());
 
-    match conf.import(CONFIG_FILE) {
-        Ok(_) => println!("Configuration has been loaded successfully!"),
-        Err(e) => {
-            println!("Oops, failed loading configurations: {}", e);
-            exit(1);
-        }
-    };
-    run(conf)
+    run()
 }
