--- crates/config/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "config"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-authors.workspace = true
-
-[dependencies]
-clap = "4.6.0"
-dotenvy = "0.15.7"
-get_fields = "0.1.0"
-serde.workspace = true
-toml.workspace = true
