+++ crates/report/Cargo.toml
@@ -17,7 +17,6 @@ libsystemd-sys = "0.9.4"
 systemd = "0.10.1"
 dbus = { workspace = true }
 utils = { path = "../utils" }
-config = { path = "../config" }
 # cli = { path = "../cli" }
 dbus-crossroads = { workspace = true }
 nom = { workspace = true }
