+++ crates/utils/src/lib.rs
@@ -1,3 +1,4 @@
 #![allow(dead_code)]
 
+pub mod config;
 pub mod notify;
